# Improve MySQLnd quote escaping performance