Closes GH-13440

Commits are split for easier reviewing, and I would recommend to review them in order one by one.
I would commit them one by one separately upon approval.

For this script:
```php
$pdo = new PDO("mysql:host=$hostname", $username, $password);
for ($i = 0; $i < 1000 * 1000; $i++) {
    $pdo->quote('this is a test with a relatively short \"string\"...');
}
```

```
Benchmark 1: ../php-src-multitasking/sapi/cli/php_old gh13440.php
  Time (mean ± σ):     313.3 ms ±   3.3 ms    [User: 311.3 ms, System: 1.4 ms]
  Range (min … max):   310.4 ms … 321.6 ms    10 runs
 
Benchmark 2: ../php-src-multitasking/sapi/cli/php gh13440.php
  Time (mean ± σ):     129.7 ms ±   2.0 ms    [User: 125.5 ms, System: 3.7 ms]
  Range (min … max):   127.5 ms … 135.7 ms    21 runs
 
Summary
  ../php-src-multitasking/sapi/cli/php gh13440.php ran
    2.42 ± 0.05 times faster than ../php-src-multitasking/sapi/cli/php_old gh13440.php
```

For this script:
```php
$pdo = new PDO("mysql:host=$hostname", $username, $password);
$str = str_repeat('\xff', 30);
for ($i = 0; $i < 1000 * 1000; $i++) {
    $pdo->quote($str);
}
```

```
Benchmark 1: ../php-src-multitasking/sapi/cli/php_old gh13440.php
  Time (mean ± σ):     700.5 ms ±   5.5 ms    [User: 695.4 ms, System: 3.9 ms]
  Range (min … max):   692.6 ms … 713.9 ms    10 runs
 
Benchmark 2: ../php-src-multitasking/sapi/cli/php gh13440.php
  Time (mean ± σ):     229.4 ms ±   2.6 ms    [User: 224.1 ms, System: 4.6 ms]
  Range (min … max):   225.8 ms … 235.6 ms    12 runs
 
Summary
  ../php-src-multitasking/sapi/cli/php gh13440.php ran
    3.05 ± 0.04 times faster than ../php-src-multitasking/sapi/cli/php_old gh13440.php
```