@kamil-tekiela 2024-02-22
> really cool, thank you sooo much. do you think we can/could/should do similar things for
> 
> ```
>         for ($i = 0; $i < 1000 * 1000; $i++) {
>             mysqli_real_escape_string('this is a test with a relatively short \"string\"...');
>         }
> ```
> 
> ?

Probably the PR's name is a bit misleading. If PDO is using mysqlnd then it's the same code. If PDO uses libmysql then we have no control over how the escaping function works.