@kamil-tekiela 2024-02-22
One thing I just cannot understand. In the function declaration it says:
```
static unsigned int mysqlnd_mbcharlen_gb18030(const unsigned int c)
```
but we are not passing an integer to it, right? How does that work?