@nielsdos 2024-02-22
> For example, https://github.com/mysql/mysql-server/blob/trunk/mysys/charset.cc#L439 but I still do not understand how it works. Maybe it will help you?

So it looks like we _can_ hit that, but we don't actually seem to test that in PHP's test suite. I read the code you linked and I think it clicked now.

I'm going to reuse their example, but I'll try to explain it with more words why [the character length check](https://github.com/mysql/mysql-server/blob/824e2b4064053f7daf17d7f3f84b7a3ed92e5fb4/mysys/charset.cc#L443-L445) is there.

Assume we are using the GBK encoding.
Assume our input is `0xbf 0x27`.
Assume we don't do the character length check, so we can see what would happen.

1. First iteration, i.e. `*from == 0xbf`. This isn't a valid GBK multibyte sequence start, so the validity check here fails: https://github.com/mysql/mysql-server/blob/824e2b4064053f7daf17d7f3f84b7a3ed92e5fb4/mysys/charset.cc#L423
Without the character length check, we'd check if we need to escape the current character `0xbf`. The character `0xbf` isn't handled in the switch case so we don't escape it and append `0xbf` to the output buffer.
2. Second iteration, i.e. `*from == 0x27`. This isn't a valid start either, so we go to the escape logic. Note that `0x27` is the character `'`, so we have to escape! We write two bytes to the output: `\` (this is `0x5c`) and `'` (this is `0x27`).
3. The function finished, let's look at the output: `0xbf 0x5c 0x27`. Now we actually made a problem: `0xbf 0x5c` is a valid GBK multibyte sequence! So we transformed an invalid multibyte sequences into a valid one, potentially corrupting data. The solution is to check whether it could have been part of a multibyte sequence, but the checks are less strict.

----

> but we are not passing an integer to it, right? How does that work?

I think this is just an API oddity. Looking at the mysql code:

https://github.com/mysql/mysql-server/blob/824e2b4064053f7daf17d7f3f84b7a3ed92e5fb4/strings/ctype.cc#L972

this passes only the current byte to `my_mbcharlen`, which is similar what we do.