@kamil-tekiela 2024-02-22
Thank you. Your explanation actually made it clear to me. I tested right now and it works exactly how you described.