@nielsdos 2024-02-23
> Could you rebase the unit test onto master once the tests pass, please? It should be committed regardless of this PR. Also maybe change the description of the test to reflect that it tests escaping with GBK.

Sure, I'll do that.