@kamil-tekiela 2024-02-23
Please correct me if I am wrong. The optimization assumes that if the byte is within the range 0-127 and the character set is multibyte with minimum length of 1 character, then it is safe to treat this byte as ASCII? 

Since we always check the first byte of the multibyte sequence, csets such as SJIS are not a problem. https://en.wikipedia.org/wiki/Shift_JIS#Structure

The problem would only be if a cset uses one of the 128 bytes as a starting byte in a multibyte sequence. For example, [ISO-2022-JP](https://en.wikipedia.org/wiki/ISO-2022-JP) would cause a problem, but afaik MySQL doesn't support it.