@nielsdos 2024-02-24
That is indeed correct.
The commit itself has a commit description where I checked whether that holds for all encodings.