@nielsdos 2024-02-24
> I guess it is ok to merge. I really hope that we haven't missed anything as this is a very important function in terms of security.

The overflow check -> assert change should be fine because each byte can at most output 2 bytes, and the truncation change seems safe too.
Indeed the change to avoid the call to mb_valid is trickier. I could fuzz it for a while, i.e. write a fuzzer harness that tries to find inputs that result in a different output when given to the unoptimized vs optimized version.