@staabm 2024-03-04
thanks again for working on my issue.

> The overflow check -> assert change should be fine because each byte can at most output 2 bytes, and the truncation change seems safe too.

would it make sense to merge the commits which are less risky, so some low hanging fruits are not blocked on finding time to write the fuzzer?