@nielsdos 2024-03-04
I noticed that the function is very sensitive to any modification, particularly for register spills and possibly CPU performance tricks.
Removing the overflow checks was obly beneficial after also guarding the valid check.
I'll look into fuzzing this as a whole sometime soon.