@nielsdos 2024-03-04
I don't see libmysqlclient doing much special in comparison to us.
I'm going to think a bit about possible compiler tricks to reduce the impact of the indirect call instead of using a potential dangerous patch. I think it's possible because the compiler generates pessimistic code because the call target "could go to anywhere" in the eyes of the compiler. We know where it may go though, and by somehow telling that to the compiler we may be able to get interprocedural optimizations and optimistic code.