@nielsdos 2024-03-06
I've pushed two commits:
* The first one undoes the dangerous optimization; and replaces it with a purely compiler-trick-based optimization. This results in a runtime average of `250.9 ms` which is 1.25x faster than the original. The idea here is to remove all the virtual calls and replace it by a jump table (due to the switch), resulting in much better code generation. Instead of storing function pointers in the charset data, I store an enum member. A dispatch function will then call the right helper based on what the enum value was.
* The second one retries the dangerous optimization, but in a safer way. To make sure I introduced no bugs I have added an (optional) sanity check that exhaustively tests inputs to check for inconsistencies. The idea here is to store for each charset what the lowest number is that a multibyte sequence can start with. For typical ASCII-compatible charsets this is 0x80, and for some others it is even higher than that (i.e. 0xA1). For complex charsets, or charsets that always emit more than one byte it is 0x00 (because there's always a multibyte sequence). We use this trick to skip the valid & mbcharlen calls whereever possible. This is the biggest save of the two commits.

These two combined result in a runtime even better than the original PR had:
```
  Time (mean ± σ):     120.2 ms ±   1.2 ms    [User: 116.9 ms, System: 2.8 ms]
  Range (min … max):   118.0 ms … 122.9 ms    24 runs
```